# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  staff: 'staff',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
}

export interface AuthResponse {
  user: User;
}

export type ParkingSlotSlotType = typeof ParkingSlotSlotType[keyof typeof ParkingSlotSlotType];


export const ParkingSlotSlotType = {
  car: 'car',
  bike: 'bike',
  truck: 'truck',
} as const;

export type ParkingSlotStatus = typeof ParkingSlotStatus[keyof typeof ParkingSlotStatus];


export const ParkingSlotStatus = {
  available: 'available',
  occupied: 'occupied',
} as const;

export interface ParkingSlot {
  id: number;
  slotNumber: string;
  slotType: ParkingSlotSlotType;
  status: ParkingSlotStatus;
  createdAt?: string;
}

export type SlotInputSlotType = typeof SlotInputSlotType[keyof typeof SlotInputSlotType];


export const SlotInputSlotType = {
  car: 'car',
  bike: 'bike',
  truck: 'truck',
} as const;

export interface SlotInput {
  slotNumber: string;
  slotType: SlotInputSlotType;
}

export type SlotUpdateSlotType = typeof SlotUpdateSlotType[keyof typeof SlotUpdateSlotType];


export const SlotUpdateSlotType = {
  car: 'car',
  bike: 'bike',
  truck: 'truck',
} as const;

export type SlotUpdateStatus = typeof SlotUpdateStatus[keyof typeof SlotUpdateStatus];


export const SlotUpdateStatus = {
  available: 'available',
  occupied: 'occupied',
} as const;

export interface SlotUpdate {
  slotNumber?: string;
  slotType?: SlotUpdateSlotType;
  status?: SlotUpdateStatus;
}

export type VehicleEntryVehicleType = typeof VehicleEntryVehicleType[keyof typeof VehicleEntryVehicleType];


export const VehicleEntryVehicleType = {
  car: 'car',
  bike: 'bike',
  truck: 'truck',
} as const;

export interface VehicleEntry {
  id: number;
  vehicleNumber: string;
  ownerName: string;
  vehicleType: VehicleEntryVehicleType;
  slotId: number;
  /** @nullable */
  slotNumber?: string | null;
  entryTime: string;
}

export type VehicleEntryInputVehicleType = typeof VehicleEntryInputVehicleType[keyof typeof VehicleEntryInputVehicleType];


export const VehicleEntryInputVehicleType = {
  car: 'car',
  bike: 'bike',
  truck: 'truck',
} as const;

export interface VehicleEntryInput {
  vehicleNumber: string;
  ownerName: string;
  vehicleType: VehicleEntryInputVehicleType;
}

export interface VehicleExitInput {
  vehicleNumber: string;
}

export interface ParkingTransaction {
  id: number;
  vehicleNumber: string;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  vehicleType?: string | null;
  slotId: number;
  /** @nullable */
  slotNumber?: string | null;
  entryTime: string;
  exitTime: string;
  durationMinutes: number;
  fee: number;
}

export interface DashboardStats {
  totalSlots: number;
  availableSlots: number;
  occupiedSlots: number;
  todayRevenue: number;
  todayVehicles: number;
  occupancyRate: number;
  weekRevenue?: number;
  monthRevenue?: number;
}

export interface OccupancyDataPoint {
  hour: string;
  count: number;
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
}

export interface Report {
  period: string;
  totalVehicles: number;
  totalRevenue: number;
  averageFee?: number;
  transactions: ParkingTransaction[];
}

export interface ExportData {
  csv: string;
  filename: string;
}

export type ListSlotsParams = {
status?: ListSlotsStatus;
type?: string;
};

export type ListSlotsStatus = typeof ListSlotsStatus[keyof typeof ListSlotsStatus];


export const ListSlotsStatus = {
  available: 'available',
  occupied: 'occupied',
} as const;

export type ListActiveVehiclesParams = {
search?: string;
};

export type SearchVehicleParams = {
vehicleNumber: string;
};

export type ListTransactionsParams = {
search?: string;
startDate?: string;
endDate?: string;
limit?: number;
offset?: number;
};

export type GetOccupancyChartParams = {
date?: string;
};

export type GetRevenueChartParams = {
period?: GetRevenueChartPeriod;
};

export type GetRevenueChartPeriod = typeof GetRevenueChartPeriod[keyof typeof GetRevenueChartPeriod];


export const GetRevenueChartPeriod = {
  week: 'week',
  month: 'month',
} as const;

export type GetDailyReportParams = {
date?: string;
};

export type GetWeeklyReportParams = {
startDate?: string;
};

export type GetMonthlyReportParams = {
month?: number;
year?: number;
};

export type ExportReportParams = {
type: ExportReportType;
date?: string;
};

export type ExportReportType = typeof ExportReportType[keyof typeof ExportReportType];


export const ExportReportType = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;