# lib/api-zod/src/generated/types/parkingTransaction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ParkingTransaction {
  id: number;
  vehicleNumber: string;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  vehicleType?: string | null;
  slotId: number;
  /** @nullable */
  slotNumber?: string | null;
  entryTime: string;
  exitTime: string;
  durationMinutes: number;
  fee: number;
}